requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
configparser==3.5.0
cryptography==2.3.1
enum34==1.1.6
future==0.16.0
gevent==1.3.6
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.7
ipaddress==1.0.22
pkg-resources==0.0.0
pycparser==2.19
pyguacamole==0.6
six==1.11.0
